Don't run git command as root

+++ server/lib/git.coffee
@@ -3,6 +3,7 @@ request = require 'request'
 compareVersions = require 'mozilla-version-comparator'
 exec = require('child_process').exec
 executeUntilEmpty = require '../helpers/executeUntilEmpty'
+directory = require './directory'
 conf = require('./conf').get
 log = require('printit')
     prefix: 'lib:git'
@@ -99,37 +100,48 @@ module.exports.init = (app, callback) ->
         * Update submodule
 ###
 module.exports.update = (app, callback) ->
+    directory.changeOwner app.user, app.dir, (err) ->
+        if err
+            log.error err
+            callback err
 
-    # Default branch is master
-    # branch can store in app.repository (controller manifest) or app.branch (database)
-    branch = app.repository.branch or app.branch or "master"
+        else
+            # Default branch is master
+            # branch can be stored in app.repository (controller manifest)
+            # or app.branch (database)
+            branch = app.repository.branch or app.branch or "master"
 
-    # Setup the git commands to be executed
-    commands = [
-        "git reset --hard "
-        "git pull origin #{branch}"
-        "git submodule update --recursive"
-    ]
+            # Setup the git commands to be executed
+            commands = [
+                "git reset --hard "
+                "git pull origin #{branch}"
+                "git submodule update --recursive"
+            ]
 
-    config =
-        cwd: app.dir # runs all the command in the app's directory
-        env: "USER": app.user
+            config =
+                cwd: app.dir # runs all the command in the app's directory
+                user: app.user
 
-    executeUntilEmpty commands, config, callback
+            executeUntilEmpty commands, config, callback
 
 ###
     Change branch of <app>
 ###
 module.exports.changeBranch = (app, newBranch, callback) ->
+    directory.changeOwner app.user, app.dir, (err) ->
+        if err
+            log.error err
+            callback err
 
-    # Setup the git commands to be executed
-    commands = [
-        "git fetch origin #{newBranch}:#{newBranch}"
-        "git checkout #{newBranch}"
-    ]
+        else
+            # Setup the git commands to be executed
+            commands = [
+                "git fetch origin #{newBranch}:#{newBranch}"
+                "git checkout #{newBranch}"
+            ]
 
-    config =
-        cwd: app.dir # runs all the command in the app's directory
-        env: "USER": app.user
+            config =
+                cwd: app.dir # runs all the command in the app's directory
+                user: app.user
 
-    executeUntilEmpty commands, config, callback
+            executeUntilEmpty commands, config, callback
